// src/types.ts
export type Department =
  | "civil"
  | "electrical"
  | "mechanical"
  | "computing"
  | "biotechnology";

export type Page =
  | "home"
  | "about"
  | "departments"
  | "programmes"
  | "announcements"
  | "events"
  | "calendar"
  | "feedback"
  | "contact"
  | "gallery"
  | "profile"
  | "login";